simple_3d_shapes mutated into a music visualizer

+++ protowizard/protographics.h
@@ -77,7 +77,7 @@ public:
 
 	static ProtoGraphics* getInstance() { return instance; }
 	
-	float klock();
+	double klock();
 
 	int getMouseX();
